Improve settings UI
