+++ src/views/settings/DebugSettings.ts
@@ -1,45 +1,20 @@
 import { ButtonComponent, Setting } from "obsidian";
 import ObsidianHardcover from "src/main";
-import { addSyncButton } from "./SyncSettings";
+import { renderSyncButton } from "./SyncSettings";
 
-export function renderDebugInfo(
+export function renderDebugSection(
 	containerEl: HTMLElement,
 	plugin: ObsidianHardcover,
 	debugBookLimit: number,
 	onDebugLimitChanged: (limit: number) => void,
 	onSyncComplete: () => void,
 ): ButtonComponent {
-	const debugDetails = containerEl.createEl("details", {
-		cls: "obhc-debug-details",
-	});
+	new Setting(containerEl).setDesc(
+		`Hardcover User ID: ${plugin.settings.userId || "Not set"} - Total Hardcover Books Count: ${plugin.settings.booksCount || "Unknown"}`,
+	);
 
-	debugDetails.createEl("summary", {
-		text: "Debug",
-		cls: "obhc-debug-summary",
-	});
-
-	const debugContent = debugDetails.createEl("div", {
-		cls: "obhc-debug-content",
-	});
-
-	const infoContainer = debugContent.createEl("div", {
-		cls: "obhc-debug-info-container",
-	});
-
-	infoContainer.createEl("div", {
-		text: `Hardcover User ID: ${plugin.settings.userId || "Not set"}`,
-		cls: "obhc-debug-info-item",
-	});
-
-	infoContainer.createEl("div", {
-		text: `Total Hardcover Books Count: ${
-			plugin.settings.booksCount || "Unknown"
-		}`,
-		cls: "obhc-debug-info-item",
-	});
-
-	const button = addSyncButton({
-		containerEl: debugContent,
+	return renderSyncButton({
+		containerEl: containerEl,
 		plugin: plugin,
 		name: "Test sync",
 		description:
@@ -51,29 +26,4 @@ export function renderDebugInfo(
 		onDebugLimitChanged: onDebugLimitChanged,
 		onSyncComplete: onSyncComplete,
 	});
-
-	return button;
-}
-
-export function renderDevOptions(
-	containerEl: HTMLElement,
-	plugin: ObsidianHardcover,
-): void {
-	new Setting(containerEl)
-		.setName("Reset plugin")
-		.setDesc("Reset plugin settings to defaults")
-		.addButton((button) => {
-			button
-				.setButtonText("Reset settings")
-				.setWarning()
-				.onClick(async () => {
-					const confirmed = confirm(
-						"Are you sure you want to reset all plugin settings to defaults? This cannot be undone.",
-					);
-
-					if (confirmed) {
-						await plugin.resetSettings();
-					}
-				});
-		});
 }
